+++ backend/app/video/views.py
@@ -1,10 +1,10 @@
+from rest_framework import generics, status
+from rest_framework.response import Response
+
 from app.models import Video, VideoGroup, VideoGroupMember
+from app.utils.decorators import authenticated_view_with_error_handling
 from app.utils.mixins import AuthenticatedViewMixin, DynamicSerializerMixin
 from app.utils.responses import create_error_response
-from rest_framework import generics, status
-from rest_framework.decorators import api_view, permission_classes
-from rest_framework.permissions import IsAuthenticated
-from rest_framework.response import Response
 
 from .serializers import (VideoCreateSerializer, VideoGroupCreateSerializer,
                           VideoGroupDetailSerializer, VideoGroupListSerializer,
@@ -215,40 +215,7 @@ def _check_and_get_member(
     return member, None
 
 
-# DRY原則: 認証必須APIビューデコレーター
-def authenticated_api_view(methods):
-    """認証必須のAPIビューデコレーター（DRY原則）"""
-
-    def decorator(view_func):
-        # utils/mixins.pyのAuthenticatedViewMixinと同様の設定を使用
-        return permission_classes(AuthenticatedViewMixin.permission_classes)(
-            api_view(methods)(view_func)
-        )
-
-    return decorator
-
-
-def with_error_handling(view_func):
-    """共通のエラーハンドリングデコレーター（DRY原則）"""
-
-    def wrapper(*args, **kwargs):
-        try:
-            return view_func(*args, **kwargs)
-        except Exception as e:
-            return create_error_response(str(e), status.HTTP_500_INTERNAL_SERVER_ERROR)
-
-    return wrapper
-
-
-def authenticated_view_with_error_handling(methods):
-    """認証とエラーハンドリングを組み合わせたデコレーター（DRY原則）"""
-
-    def decorator(view_func):
-        # エラーハンドリングを最初に適用し、次に認証を適用
-        wrapped = with_error_handling(view_func)
-        return authenticated_api_view(methods)(wrapped)
-
-    return decorator
+# DRY原則: 共通デコレーターはapp.utils.decoratorsからインポート済み
 
 
 @authenticated_view_with_error_handling(["POST"])
@@ -366,7 +333,7 @@ def reorder_videos_in_group(request, group_id):
 
     # リクエストボディからvideo_idsの配列を取得
     try:
-        video_ids = request.data.get('video_ids', [])
+        video_ids = request.data.get("video_ids", [])
         if not isinstance(video_ids, list):
             return create_error_response(
                 "video_idsは配列である必要があります", status.HTTP_400_BAD_REQUEST
@@ -378,29 +345,28 @@ def reorder_videos_in_group(request, group_id):
 
     # グループ内の動画メンバーを取得（N+1問題対策）
     # select_relatedでvideoデータも事前取得
-    members = VideoGroupMember.objects.filter(group=group).select_related('video')
-    
+    members = VideoGroupMember.objects.filter(group=group).select_related("video")
+
     # 指定されたvideo_idsがグループ内の動画と一致するかチェック
     # O(1)ルックアップのためにSetを使用
     group_video_ids = set(member.video_id for member in members)
     if set(video_ids) != group_video_ids:
         return create_error_response(
-            "指定された動画IDがグループ内の動画と一致しません", status.HTTP_400_BAD_REQUEST
+            "指定された動画IDがグループ内の動画と一致しません",
+            status.HTTP_400_BAD_REQUEST,
         )
 
     # 順序を更新（N+1問題対策）
     # bulk_updateを使用して一括更新
     member_dict = {member.video_id: member for member in members}
     members_to_update = []
-    
+
     for index, video_id in enumerate(video_ids):
         member = member_dict[video_id]
         member.order = index
         members_to_update.append(member)
-    
-    # 一括更新でN+1問題を解決
-    VideoGroupMember.objects.bulk_update(members_to_update, ['order'])
 
-    return Response(
-        {"message": "動画の順序を更新しました"}, status=status.HTTP_200_OK
-    )
+    # 一括更新でN+1問題を解決
+    VideoGroupMember.objects.bulk_update(members_to_update, ["order"])
+
+    return Response({"message": "動画の順序を更新しました"}, status=status.HTTP_200_OK)
